.github/workflows/dts_patient_availability-dts_decibel_v3 2/requirements.txt
anyio==3.6.1
asgiref==3.5.2
boto3==1.24.39
botocore==1.27.39
certifi==2022.12.07
charset-normalizer==2.1.0
click==8.1.3
cloudpathlib==0.9.0
greenlet==1.1.2
h11==0.13.0
idna==3.3
jmespath==1.0.1
numpy==1.23.1
pandas==1.5.0
pip==22.2.1
polars==0.18.15
polling2==0.5.0
psycopg2==2.9.3
pyarrow==13.0.0
#pydantic==1.9.1
python-dateutil==2.8.2
pytz==2022.1
PyYAML==6.0
requests==2.28.1
scikit-learn==1.2.2
scipy==1.10.1
s3transfer==0.6.0
setuptools==65.5.1
six==1.16.0
smart-open==6.0.0
sniffio==1.2.0
SQLAlchemy==1.4.39
starlette==0.25.0
typing-extensions==4.3.0
urllib3==1.26.11
uvicorn==0.17.6
wheel==0.38.0
fastparquet==2022.11.0
s3fs==0.4.2
xgboost==2.0.3
xgbse==0.2.3
xlsxwriter==3.1.9
great_expectations==0.18.10
fastapi==0.93.0
pydantic==1.10.13
mlflow==2.7.0
seaborn
connectorx